Fix remaining Kotlin compilation errors
- EmulatorActivity: change val _0 from Short to Int (it was being passed
  to Emulator.key_event which expects int, not short)
- VirtualControlEdit: add @OptIn(ExperimentalMaterial3Api::class) to
  VirtualControlEditScreen because FloatingActionButton, ModalBottomSheet
  and rememberModalBottomSheetState are experimental APIs
- VirtualControlEdit: replace .scale property access with .get_scale()
  direct method call (Kotlin was not synthesizing the property on the
  Java interface VirtualControl.Component through the safe-call chain)
- EmulatorSettingsScreen: use aenu.emulator.Emulator.Config (the nested
  Config class lives on the base Emulator class, not on aenu.ax360e.Emulator
  subclass) - import as NativeEmulator alias to avoid clash with
  aenu.ax360e.Emulator
- EmulatorSettingsScreen: handle nullability of config (it can be null
  when the config file doesn't exist or fails to open) - use ?. safe call
  for all config.load_config_entry / save_config_entry / close_config_file
  invocations
- EmulatorSettingsScreen: add missing 'is SettingsEntry.StrLeaf' branch
  to the when() expression to make it exhaustive, with a StrLeafRow
  composable that shows the path value and allows editing via a dialog

+++ app/src/main/java/aenu/ax360e/EmulatorActivity.kt
@@ -263,7 +263,7 @@ class EmulatorActivity : ComponentActivity(), SurfaceHolder.Callback, View.OnGen
             val laxisY = event.getAxisValue(MotionEvent.AXIS_Y)
             val raxisX = event.getAxisValue(MotionEvent.AXIS_Z)
             val raxisY = event.getAxisValue(MotionEvent.AXIS_RZ)
-            val _0: Short = 0
+            val _0: Int = 0
 
             // Left joystick
             if (laxisX != 0f) {
